divided tests into two modules
