+++ test/SdkUnitTests.js
@@ -25,7 +25,6 @@ var basePath = restApi.BasePath.DEMO;
 var oAuthBasePath = oAuth.BasePath.DEMO;
 
 var SignTest1File = 'docs/SignTest1.pdf';
-var SignTest2File = 'docs/SignTest1.docx';
 var LargeTestDocument1 = 'docs/LargeTestDocument1.pdf';
 var brandLogoPath = 'img/docusign-lgo.png';
 var brandXmlPath = 'docs/brand.xml';
@@ -1283,310 +1282,4 @@ describe('SDK Unit Tests:', function (done) {
         }
       });
   });
-
-  it("Get envelope by id", function (done) {
-    var envelopesApi = new docusign.EnvelopesApi(apiClient);
-
-    envelopesApi.getEnvelope(accountId, envelopeId)
-      .then(function (envelope) {
-        assert.notStrictEqual(envelope, undefined);
-        done();
-      })
-      .catch(function (error) {
-        if (error) {
-          return done(error);
-        }
-      });
-  });
-  
-  it('Get envelope recipients', function(done) {
-    var envelopesApi = new docusign.EnvelopesApi(apiClient);
-
-    envelopesApi.listRecipients(accountId, envelopeId)
-      .then(function (recipients) {
-        assert.notStrictEqual(recipients, undefined);
-        done();
-      })
-      .catch(function (error) {
-        if (error) {
-          return done(error);
-        }
-      });      
-  })
-
-  it('Create sender view', function(done) {
-    var fileBytes = null;
-    try {
-      // read file from a local directory
-      fileBytes = fs.readFileSync(path.resolve(__dirname, SignTest1File));
-    } catch (ex) {
-      // handle error
-      console.log(`Exception: ${ex}`);
-    }
-
-    // create an envelope to be signed
-    var envDef = new docusign.EnvelopeDefinition();
-    envDef.emailSubject = 'Please Sign my Node SDK Envelope';
-    envDef.emailBlurb = 'Hello, Please sign my Node SDK Envelope.';
-
-    // add a document to the envelope
-    var doc = new docusign.Document();
-    var base64Doc = Buffer.from(fileBytes).toString('base64');
-    doc.documentBase64 = base64Doc;
-    doc.name = 'TestFile.pdf';
-    doc.documentId = '1';
-
-    var docs = [];
-    docs.push(doc);
-    envDef.documents = docs;
-
-    // Add a recipient to sign the document
-    var signer = new docusign.Signer();
-    signer.email = userName;
-    var name = 'Pat Developer';
-    signer.name = name;
-    signer.recipientId = '1';
-
-    // this value represents the client's unique identifier for the signer
-    var clientUserId = '2939';
-    signer.clientUserId = clientUserId;
-
-    // create a signHere tab somewhere on the document for the signer to sign
-    // default unit of measurement is pixels, can be mms, cms, inches also
-    var signHere = new docusign.SignHere();
-    signHere.documentId = '1';
-    signHere.pageNumber = '1';
-    signHere.recipientId = '1';
-    signHere.xPosition = '100';
-    signHere.yPosition = '100';
-
-    // can have multiple tabs, so need to add to envelope as a single element list
-    var signHereTabs = [];
-    signHereTabs.push(signHere);
-    var tabs = new docusign.Tabs();
-    tabs.signHereTabs = signHereTabs;
-    signer.tabs = tabs;
-
-    // Above causes issue
-    envDef.recipients = new docusign.Recipients();
-    envDef.recipients.signers = [];
-    envDef.recipients.signers.push(signer);
-
-    // make the envelope with "created" (draft) status
-    envDef.status = 'created';
-
-    var envelopesApi = new docusign.EnvelopesApi(apiClient);
-
-    envelopesApi.createEnvelope(accountId, { envelopeDefinition: envDef })
-      .then(function (envelopeSummary) {
-          var returnUrl = 'http://www.docusign.com/developer-center';
-          var returnUrlRequest = new docusign.ReturnUrlRequest();
-          returnUrlRequest.returnUrl = returnUrl;
-
-          return envelopesApi.createSenderView(accountId, envelopeSummary.envelopeId, { returnUrlRequest })
-      })
-      .then(function (viewUrl) {
-        if (viewUrl) {
-          console.log(`ViewUrl is ${JSON.stringify(viewUrl)}`);
-          done();
-        }
-      })
-      .catch(function (error) {
-        if (error) {
-          return done(error);
-        }
-      });
-  })
-
-  it('Get account info', function(done) {
-    var accountsApi = new docusign.AccountsApi(apiClient);
-
-    accountsApi.getAccountInformation(accountId, { includeAccountSettings: true })
-      .then(function (accountInfo) {
-        assert.notStrictEqual(accountInfo, undefined);
-        assert.notStrictEqual(accountInfo.accountSettings, undefined);
-        done();
-      })
-      .catch(function (error) {
-        if (error) {
-          return done(error);
-        }
-      });
-  })
-
-  it('Update recipients', function(done) {
-    var newSigner = new docusign.Signer();
-    newSigner.email = userName;
-    newSigner.name = 'Signer2';
-    newSigner.recipientId = '2';
-
-    // this value represents the client's unique identifier for the signer
-    var clientUserId = '2939';
-    newSigner.clientUserId = clientUserId;
-
-    // create a signHere tab somewhere on the document for the signer to sign
-    // default unit of measurement is pixels, can be mms, cms, inches also
-    var signHere = new docusign.SignHere();
-    signHere.documentId = '1';
-    signHere.pageNumber = '1';
-    signHere.recipientId = '1';
-    signHere.xPosition = '100';
-    signHere.yPosition = '100';
-
-    // can have multiple tabs, so need to add to envelope as a single element list
-    var signHereTabs = [];
-    signHereTabs.push(signHere);
-    var tabs = new docusign.Tabs();
-    tabs.signHereTabs = signHereTabs;
-    newSigner.tabs = tabs;
-
-    var newRecipients = new docusign.Recipients();
-    newRecipients.signers = [];
-    newRecipients.signers.push(newSigner);
-
-    var envelopesApi = new docusign.EnvelopesApi(apiClient);
-
-    envelopesApi.updateRecipients(accountId, envelopeId, { recipients: newRecipients })
-      .then(function (recipientsUpdateSummary) {
-        assert.notStrictEqual(recipientsUpdateSummary, undefined);
-        assert.notStrictEqual(recipientsUpdateSummary.recipientUpdateResults, undefined);
-
-        return envelopesApi.listRecipients(accountId, envelopeId)
-      })
-      .then(function (recipients) {
-        assert.notStrictEqual(recipients, undefined);
-        assert.equal(recipients.recipientCount, 2);
-        done();
-      })
-      .catch(function (error) {
-        if(error){
-          return done(error);
-        }
-      });
-  })
-
-  it('Get templates', function(done) {
-    var templatesApi = new docusign.TemplatesApi(apiClient);
-
-    templatesApi.listTemplates(accountId)
-      .then(function (templateResults) {
-        assert.notStrictEqual(templateResults, undefined);
-        assert.notStrictEqual(templateResults.envelopeTemplates, undefined);
-        assert.notStrictEqual(templateResults.envelopeTemplates[0], undefined);
-        done();
-      })
-      .catch(function (error) {
-        if (error) {
-          return done(error);
-        }
-      });
-  })
-
-  it('Get users', function(done) {
-    var usersApi = new docusign.UsersApi(apiClient);
-
-    usersApi.list(accountId)
-      .then(function (userInformationList) {
-        assert.notStrictEqual(userInformationList, undefined);
-        assert.notStrictEqual(userInformationList.users, undefined);
-        assert.notStrictEqual(userInformationList.users[0], undefined);
-        done();
-      })
-      .catch(function (error) {
-        if (error) {
-          return done(error);
-        }
-      });
-  })
-
-  it('Get audit events', function(done) {
-    var envelopesApi = new docusign.EnvelopesApi(apiClient);
-
-    envelopesApi.listAuditEvents(accountId, envelopeId)
-      .then(function (envelopeAuditEventResponse) {
-        assert.notStrictEqual(envelopeAuditEventResponse, undefined);
-        assert.notStrictEqual(envelopeAuditEventResponse.auditEvents, undefined);
-        assert.notStrictEqual(envelopeAuditEventResponse.auditEvents[0], undefined);
-        done();
-      })
-      .catch(function (error) {
-        if (error) {
-          return done(error);
-        }
-      });
-  })
-
-  it('Update documents', function(done) {
-    var newFileBytes = null;
-    try {
-      // read file from a local directory
-      newFileBytes = fs.readFileSync(path.resolve(__dirname, SignTest2File));
-    } catch (ex) {
-      // handle error
-      console.log(`Exception: ${ex}`);
-    }
-
-    var envelopesApi = new docusign.EnvelopesApi(apiClient);
-
-    var newEnvDef = new docusign.EnvelopeDefinition();
-    
-    // add a document to the envelope
-    var newDoc = new docusign.Document();
-    var base64Doc2 = Buffer.from(newFileBytes).toString('base64');
-    newDoc.documentBase64 = base64Doc2;
-    newDoc.name = 'TestFile.docx';
-    newDoc.documentId = '2';
-
-    var newDocs = [];
-    newDocs.push(newDoc);
-    newEnvDef.documents = newDocs;
-    var oldDocumentsCount = 0;
-
-    envelopesApi.listDocuments(accountId, envelopeId)
-      .then(function (oldDocuments) {
-        assert.notStrictEqual(oldDocuments, undefined);
-        assert.notStrictEqual(oldDocuments.envelopeDocuments, undefined);
-        oldDocumentsCount = oldDocuments.envelopeDocuments.length;
-
-        return envelopesApi.updateDocuments(accountId, envelopeId, { envelopeDefinition: newEnvDef })
-      })
-      .then(function (envelopeDocumentsResult) {
-        assert.notStrictEqual(envelopeDocumentsResult, undefined);
-
-        return envelopesApi.listDocuments(accountId, envelopeId)
-      })
-      .then(function (documents) {
-        assert.notStrictEqual(documents, undefined);
-        assert.notStrictEqual(documents.envelopeDocuments, undefined);
-        assert.equal(documents.envelopeDocuments.length, oldDocumentsCount + 1);
-        done();
-      })
-      .catch(function (error) {
-        if (error) {
-          return done(error);
-        }
-      });
-  })
-
-  it('Get recipient tabs', function(done) {
-    var envelopesApi = new docusign.EnvelopesApi(apiClient);
-
-    envelopesApi.listRecipients(accountId, envelopeId)
-      .then(function (recipients) {
-        assert.notStrictEqual(recipients, undefined);
-        assert.notStrictEqual(recipients.signers, undefined);
-        assert.notStrictEqual(recipients.signers[0], undefined);
-        
-        return envelopesApi.listTabs(accountId, envelopeId, recipients.signers[0].recipientId)
-      })
-      .then(function(tabs) {
-        assert.notStrictEqual(tabs, undefined);
-        done();
-      })
-      .catch(function (error) {
-        if (error) {
-          return done(error);
-        }
-      });
-  })
 });
